[WIP] broke DriverFramework.cpp into components
